requirements.txt
certifi==2022.12.7
grpcio==1.53.0
google-cloud-translate==3.11.1
python-telegram-bot==20.2
requests==2.28.2
boto3==1.26.104
